// src/spec/threatgile.generated.ts
export interface Threagile {
// ...
  individual_risk_categories?: {
    [k: string]: {
      /**
       * ID
       */
      id: string;
      /**
       * Description
       */
      description: string | null;
      /**
       * Impact
       */
      impact: string;
      /**
       * ASVS
       */
      asvs: string;
      /**
       * Cheat sheet
       */
      cheat_sheet: string;
      /**
       * Action
       */
      action: string;
      /**
       * Mitigation
       */
      mitigation: string;
      /**
       * Check
       */
      check: string;
      /**
       * Function
       */
      function: "business-side" | "architecture" | "development" | "operations";
      /**
       * STRIDE
       */
      stride:
        | "spoofing"
        | "tampering"
        | "repudiation"
        | "information-disclosure"
        | "denial-of-service"
        | "elevation-of-privilege";
      /**
       * Detection logic
       */
      detection_logic: string;
      /**
       * Risk assessment
       */
      risk_assessment: string;
      /**
       * False positives
       */
      false_positives: string;
      /**
       * Model failure possible reason
       */
      model_failure_possible_reason: boolean;
      /**
       * CWE
       */
      cwe: number;
      /**
       * Risks identified
       */
      risks_identified: {
        [k: string]: {
          /**
           * Severity
           */
          severity?: "low" | "medium" | "elevated" | "high" | "critical";
          /**
           * Exploitation likelihood
           */
          exploitation_likelihood?: "unlikely" | "likely" | "very-likely" | "frequent";
          /**
           * Exploitation impact
           */
          exploitation_impact?: "low" | "medium" | "high" | "very-high";
          /**
           * Data breach probability
           */
          data_breach_probability?: "improbable" | "possible" | "probable";
          /**
           * Data breach technical assets
           */
          data_breach_technical_assets?: string[] | null;
          /**
           * Most relevant data asset
           */
          most_relevant_data_asset?: string | null;
          /**
           * Most relevant technical asset
           */
          most_relevant_technical_asset?: string | null;
          /**
           * Most relevant communication link
           */
          most_relevant_communication_link?: string | null;
          /**
           * Most relevant trust boundary
           */
          most_relevant_trust_boundary?: string | null;
          /**
           * Most relevant shared runtime
           */
          most_relevant_shared_runtime?: string | null;
          [k: string]: unknown;
        };
      };
      [k: string]: unknown;
    };
  } | null;
// ...
}